backend/requirements.txt
openai-whisper~=20240930
sounddevice~=0.5.1
soundfile~=0.12.1
numpy~=2.0.2
Flask~=3.1.0
Flask-Migrate~=4.0.7
